apps/streamlit-dashboard/requirements.txt
streamlit==1.31.0
streamlit-webrtc==0.47.1
opencv-python==4.9.0.80
plotly==5.18.0
pandas==2.2.0
numpy==1.26.4
requests==2.31.0
websocket-client==1.7.0
Pillow==10.2.0
